frontend/src/app/home/header/header.component.ts
import { Component, OnInit } from '@angular/core';
import { SearchResultService } from '../search-result/search-result.service';
import { UserService } from 'src/app/user/shared/user.service';
import { Router } from '@angular/router';

@Component({
  selector: 'main-header',
  templateUrl: './header.component.html',
  styleUrls: ['./header.component.css']
})
export class HeaderComponent implements OnInit {

  constructor(private router: Router,private resultService: SearchResultService,private _userService: UserService) { }

  ngOnInit() {
    
  }
  onLogout(){
    this._userService.Logout();
    localStorage.removeItem('loggedIn');
    this._userService.setLoggedIn(false);
    this.router.navigate(['login']);
  }
  navigate(pagename:any){
    this.router.navigate(['profilepage'],{state:pagename});

  }
}
